+++ tests/testthat/test-keepFUN.R
@@ -1,63 +1,60 @@
-context("Keeps and counts words")
+context("KeepFUNs")
 
-str.test1 <- "Lorem ipsum dolor sit amet,"
-str.test2 <- "consectetur adipisicing elit."
+str1 <- "Lorem ipsum dolor sit amet,\n consectetur"
+str2 <- "consectetur adipisicing elit."
 
-test_that("keepWords typeof", {
-  expect_equal(typeof(keepWords(str.test1)), "list")
+res1 <- keepWords(str1)
+res2 <- keepWords(str1, collapse = NULL)
+res3a <- keepWords(c(str1, str2), slc = c(1, 3))
+res3b <- keepWords(c(str1, str2), slc = c(1, 3), collapse = NULL)
+res3c <- keepWords(c(str1, str2), slc = c(1, 3), collapse = ", ")
+res4a <- keepWords(c(str1, str2), slc = c(1, 4))
+res4b <- keepWords(c(str1, str2), slc = c(1, 4), na.rm = TRUE)
+res5a <- keepWords(str1, slc = 6)
+res5b <- keepWords(str1, slc = 6, split_words = "[[:punct:] ]")
+
+test_that("keepWords()", {
+  expect_equal(typeof(res1), "character")
+  expect_equal(typeof(res2), "list")
+  expect_equal(typeof(res3a), "character")
+  expect_equal(typeof(res3b), "list")
+  expect_identical(res3a, c("Lorem dolor", "consectetur elit"))
+  expect_identical(res3c, c("Lorem, dolor", "consectetur, elit"))
+  expect_identical(res4a, c("Lorem sit", "consectetur NA"))
+  expect_identical(res4b, c("Lorem sit", "consectetur"))
+  expect_identical(res5a, "consectetur")
+  expect_identical(res5b, "\n")
 })
 
-res1a <- keepWords(str.test1)[[1]]
-res1b <- keepWords(str.test1, slc = 5)[[1]]
-test_that("keepWords one word", {
-  expect_equal(res1a, "Lorem")
-  expect_equal(res1b, "amet")
-})
-
-res2a <- keepWords(str.test1, punct.rm = FALSE)[[1]]
-res2b <- keepWords(str.test1, punct.rm = FALSE, slc = 5)[[1]]
-test_that("keepWords keeps punctuation", {
-  expect_equal(res2a, "Lorem")
-  expect_equal(res2b, "amet,")
-})
-
-test_that("keepWords of missings", {
-  expect_equal(keepWords("")[[1]], NA_character_)
-  expect_equal(keepWords(str.test1,6)[[1]],  NA_character_)
-  expect_equal(length(keepWords("", na.rm=T)[[1]]), 0)
-})
-
-test_that("keepWords collapses", {
-  expect_equal(keepWords(str.test1, c(1,5), collapse="-")[[1]], "Lorem-amet")
-})
-
-res <- keepWords(c(str.test1,str.test2))
-test_that("keepWords for multi strings", {
-  expect_equal(length(res), 2)
-  expect_equal(length(res[[1]]), 1)
-  expect_equal(length(res[[2]]), 1)
+test_that("wordCount()", {
+  expect_equal(wordCount(c("two words!", "... and three words")), c(2, 3))
+  expect_equal(wordCount(".#\n.. "), 0)
 })
 
 
 
-##--
-str.test3 <- "consectetur-adipisicing elit."
-test_that("count words", {
-  expect_equal(wordCount(str.test1), 5)
-  expect_true(all(wordCount(c(str.test1, str.test2)) == c(5,3)))
-  expect_equal(wordCount(str.test3), 3)
-  expect_equal(wordCount(str.test3, exclude = "[ \\.]"), 2)
+res1 <- keepLetters(str1, c(1,5,6))
+res2 <- keepLetters(str1, c(1,5,6), collapse = NULL)
+res3 <- keepLetters(c(str1, str2))
+
+test_that("keepLetters()", {
+  expect_equal(keepLetters(str1), "L")
+  expect_equal(keepLetters(str1, c(1,5,6)), "Lmi")
+  expect_equal(res2, strsplit(res1, ""))
+  expect_identical(res3, c("L", "c"))
+  expect_equal(keepLetters(str1, 100), "NA")
+  expect_equal(keepLetters(str1, 100, na.rm = TRUE), "")
 })
 
 
-context("Keeps letters")
-strex <- c('Lorem ipsum', 'dolor sit', ' amet')
-res2a <- keepLetters(strex, c(1,4))
-res2b <- keepLetters(strex, c(1,4), punct.rm = TRUE)
+res1 <- keepInitials("National Basketball Association")
+res2 <- keepInitials("National Basketball Association", collapse = NULL)
+res3 <- keepInitials("National Basketball Association", collapse = "-")
+res4 <- keepInitials(c("National Basketball Association", "Creative commons"))
 
-
-test_that("keep letters", {
-  expect_true(all(res2a == c("Le", "do", " e")))
-  expect_true(all(res2b == c("Le", "do", "at")))
-  expect_warning(keepLetters(strex, c(1,6)))
+test_that("keepsInitials()", {
+  expect_equal(res1, "NBA")
+  expect_equal(res2, strsplit(res1, ""))
+  expect_equal(res3, "N-B-A")
+  expect_equal(res4, c("NBA", "Cc"))
 })