+++ tests/testthat/test-lorem.R
@@ -1,7 +1,8 @@
-context("Lorem")
+context("LoremIpsum()")
 
 
 test_that("categorize a numerial vector", {
-  expect_equal(wordCount(loremIpsum()), 70)
-  expect_equal(wordCount(loremIpsum(10)), 10)
+  expect_equal(wordCount(loremIpsum()), 201)
+  expect_equal(wordCount(loremIpsum(12)), 12)
+  expect_equal(wordCount(loremIpsum(220)), 220)
 })